Add fastfood transform support and parametrize it with dense transform

+++ src/intrinsic_dimension/util/util.py
@@ -16,9 +16,9 @@ def count_params(model):
 
 def generate_intrinsic_dimension_candidates(start, end, n_points=100, scale='log', plot=False, basedir=''):
     if scale == 'log':
-        d_values = np.logspace(start, end, num=n_points, dtype=np.uint32)
+        d_values = np.logspace(start, end, num=n_points, dtype=int)
     elif scale == 'linear':
-        d_values = np.linspace(start, end, num=n_points, dtype=np.uint32)
+        d_values = np.linspace(start, end, num=n_points, dtype=int)
     else:
         raise AttributeError(f'Unsupported scale {scale}')
